Revert "Chore/docker"
