--- .env.defaults
@@ -1,14 +1,0 @@
-DB_NAME=freenoms-domains
-PUPPETEER_HEADLESS=true
-
-DISCORD_WEBHOOK_ID=
-DISCORD_WEBHOOK_TOKEN=
-
-FREENOM_LOGIN=email
-FREENOM_PASS=password
-FREENOM_CRONJOB="0 9 * * 1"
-
-# Docker exec path
-CHROME_PATH=google-chrome-stable
-# Linux
-#CHROME_PATH=/usr/local/bin/chromium
