--- CHANGELOG.md
@@ -1,8 +1,0 @@
-# Changelog
-
-All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
-
-## [1.0.2](https://github.com/Sorok-Dva/freenom-auto-renew-domains/compare/v1.0.1...v1.0.2) (2022-08-24)
-### Features
-* Add changelog
-* dockerization ([cf829a2](https://github.com/Sorok-Dva/freenom-auto-renew-domains/commit/cf829a2bb6e1e2f56ad703d1042390dedba2cc54))
